dsfsdf

+++ backend/src/main/java/com/group/backend/dto/CartDetailDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class CartDetailDTO {
     private long id;
     private int quantity;
-    private long unitPrice;
+    private int unitPrice;
     private LaptopSummaryDTO laptop;
 }
